+++ market/src/main/java/com/charly/market/auth/service/AuthServiceImpl.java
@@ -7,12 +7,11 @@ import com.charly.market.auth.repository.AuthRedisRepository;
 import com.charly.market.global.security.JwtProperties;
 import com.charly.market.global.security.JwtTokenProvider;
 import com.charly.market.global.security.JwtTokenProvider.TokenPair;
-import com.charly.market.user.model.User;
+import com.charly.market.user.model.entity.User;
 import com.charly.market.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import java.time.Duration;
-import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
