+++ market/src/main/java/com/charly/market/user/model/entity/User.java
@@ -1,20 +1,15 @@
-package com.charly.market.user.model;
+package com.charly.market.user.model.entity;
 
 import com.charly.market.global.constant.UserRole;
 import com.charly.market.global.model.BaseTimeEntity;
-import jakarta.persistence.CollectionTable;
 import jakarta.persistence.Column;
-import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.Table;
-import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -36,7 +31,7 @@ public class User extends BaseTimeEntity {
   private Long id;
 
   @Column
-  private String username;
+  private String username; // 설명
 
   @Column(unique = true)
   private String email;
