+++ market/src/main/java/com/charly/market/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.charly.market.user.repository;
 
-import com.charly.market.user.model.User;
+import com.charly.market.user.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
