+++ market/src/main/java/com/charly/market/user/service/UserServiceImpl.java
@@ -2,7 +2,7 @@ package com.charly.market.user.service;
 
 
 import com.charly.market.global.constant.UserRole;
-import com.charly.market.user.model.User;
+import com.charly.market.user.model.entity.User;
 import com.charly.market.user.model.dto.CreateUserRequest;
 import com.charly.market.user.model.dto.UserResponse;
 import com.charly.market.user.repository.UserRepository;
